### 2024-01-24 #6278 @editorialbot
[PRE REVIEW]: EllipticForest: A Direct Solver Library for Elliptic Partial Differential Equations on Adaptive Meshes
**Submitting author:** @DamynChipman (<a href="http://orcid.org/0000-0001-6600-3720">Damyn Chipman</a>)
**Repository:** https://github.com/DamynChipman/EllipticForest
**Branch with paper.md** (empty if default branch): v1-joss-paper
**Version:** v1.0.1
