+++ paper.bib
@@ -19,7 +19,8 @@
 	pages = {A2023--A2046},
 	title = {{A direct solver with O(N) complexity for variable coefficient elliptic PDEs discretized via a high-order composite spectral collocation method}},
 	volume = {36},
-	year = {2014}}
+	year = {2014},
+	doi = {10.1137/130918988}}
 
 @book{martinsson2019fast,
 	author = {Martinsson, PG},
@@ -62,7 +63,8 @@
 	booktitle = {Visualization Handbook},
 	publisher = {Elesvier},
 	title = {{ParaView}: An End-User Tool for Large Data Visualization},
-	year = {2005}}
+	year = {2005},
+	doi = {10.1016/b978-012387582-2/50038-1}}
 
 @techreport{anl2023petsc,
 	author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed Brown and Peter Brune and Kris Buschelman and Emil Constantinescu and Lisandro Dalcin and Alp Dener and Victor Eijkhout and Jacob Faibussowitsch and William~D. Gropp and V\'{a}clav Hapla and Tobin Isaac and Pierre Jolivet and Dmitry Karpeev and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott Kruger and Dave~A. May and Lois Curfman McInnes and Richard Tran Mills and Lawrence Mitchell and Todd Munson and Jose~E. Roman and Karl Rupp and Patrick Sanan and Jason Sarich and Barry~F. Smith and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao Zhang},
@@ -83,7 +85,8 @@
 	pages = {179--203},
 	publisher = {Springer},
 	title = {Theory and application of Steklov-Poincar{\'e} operators for boundary-value problems},
-	year = {1991}}
+	year = {1991}
+	doi = {10.1007/978-94-009-1908-2_14}}
 
 @incollection{HPV:VisIt,
 	author = {Hank Childs and Eric Brugger and Brad Whitlock and Jeremy Meredith and Sean Ahern and David Pugmire and Kathleen Biagas and Mark Miller and Cyrus Harrison and Gunther H. Weber and Hari Krishnan and Thomas Fogal and Allen Sanderson and Christoph Garth and E. Wes Bethel and David Camp and Oliver R\"{u}bel and Marc Durant and Jean M. Favre and Paul Navr\'{a}til},
