+++ paper.bib
@@ -85,7 +85,7 @@
 	pages = {179--203},
 	publisher = {Springer},
 	title = {Theory and application of Steklov-Poincar{\'e} operators for boundary-value problems},
-	year = {1991}
+	year = {1991},
 	doi = {10.1007/978-94-009-1908-2_14}}
 
 @incollection{HPV:VisIt,
