### 2024-01-26 #6278 @vissarion
:wave: @asalmgren, @QibangLiu, @jeremiah-corrado, @likask, @roshansamuel would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
