### 2024-01-30 #6278 @vissarion
Hi @josemanuel22, @sandeshkatakam, @lukeolson, @kris-rowe and @chaypaterson

Would any of you be willing to review this submission for JOSS?
