### 2024-02-05 5151ed4
Continued work on adaptive rebuild.
