### 2024-02-08 #6278 @chaypaterson
How long is the article?

Chay

On Thu, 8 Feb 2024, 10:11 am Vissarion Fisikopoulos, <
***@***.***> wrote:

> Hi @mbudisic <https://github.com/mbudisic>, @slayoo
> <https://github.com/slayoo>, @dawbarton <https://github.com/dawbarton>,
> @jgoldfar <https://github.com/jgoldfar> and @arghdos.
>
> Would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
