### 2024-02-08 #6278 @vissarion
@chaypaterson thanks for your reply. The paper it is 6 pages long (you can find the pdf in the current thread) but a JOSS reviewer is expected to also review other aspects of the software package such as documentation and functionality see https://joss.readthedocs.io/en/latest/review_criteria.html for more details.
