### 2024-02-08 #6278 @lukeolson
> Hi @josemanuel22, @sandeshkatakam, @lukeolson, @kris-rowe and @chaypaterson
>
> Would any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I can do a review if you need another reviewer.
