### 2024-02-09 #6278 @vissarion
@DamynChipman with the two assigned reviewers @sandeshkatakam and @lukeolson we can move forward to the review (this will be in a new thread)

@sandeshkatakam, @lukeolson thanks for your willingness to review for JOSS!
