### 2024-02-09 a9435e0
WIP: Debugging adaptive rebuild.
