+++ src/HPSAlgorithm.hpp
@@ -477,7 +477,7 @@ public:
      * @param gamma Upper-left patch
      * @param omega Upper-right patch
      */
-    virtual void merge4to1(PatchType& tau, PatchType& alpha, PatchType& beta, PatchType& gamma, PatchType& omega) {
+    static void merge4to1(PatchType& tau, PatchType& alpha, PatchType& beta, PatchType& gamma, PatchType& omega) {
         
         EllipticForestApp& app = EllipticForestApp::getInstance();
         // app.logHead("Merging:");
@@ -504,59 +504,98 @@ public:
         // reorderOperators(tau, alpha, beta, gamma, omega);
         // mergePatch(tau, alpha, beta, gamma, omega);
 
-        // app.log("Merging:");
-        // app.log("  alpha = \n" + alpha.str());
-        // app.log("  beta = \n" + beta.str());
-        // app.log("  gamma = \n" + gamma.str());
-        // app.log("  omega = \n" + omega.str());
-        // app.log("  tau = \n" + tau.str());
+        app.log("Merging:");
+        app.log("  alpha = \n" + alpha.str());
+        app.log("  beta = \n" + beta.str());
+        app.log("  gamma = \n" + gamma.str());
+        app.log("  omega = \n" + omega.str());
+        app.log("  tau before = \n" + tau.str());
 
         // Get data
-        std::vector<PatchType*> patches = {&alpha, &beta, &gamma, &omega};
-        std::vector<PatchGridType*> grids = {&alpha.grid(), &beta.grid(), &gamma.grid(), &omega.grid()};
-        std::vector<int> sides(4);
-        std::vector<int> tags(4);
+        // std::vector<PatchGridType*> grids = {&alpha.grid(), &beta.grid(), &gamma.grid(), &omega.grid()};
+        // std::vector<int> sides(4);
+        // std::vector<int> tags(4);
         
-        // Get vector of side lengths
-        for (auto i = 0; i < 4; i++) 
-            sides[i] = patches[i]->size();
+        // // Get vector of side lengths
+        // for (auto i = 0; i < 4; i++) 
+        //     sides[i] = patches[i]->size();
 
-        // Get minimum side length
-        int min_side = *std::min_element(sides.begin(), sides.end());
+        // // Get minimum side length
+        // int min_side = *std::min_element(sides.begin(), sides.end());
 
-        // Get tags based on side lengths
-        for (auto i = 0; i < 4; i++) 
-            tags[i] = static_cast<int>(log2(sides[i])) - static_cast<int>(log2(min_side));
+        // // Get tags based on side lengths
+        // for (auto i = 0; i < 4; i++) 
+        //     tags[i] = static_cast<int>(log2(sides[i])) - static_cast<int>(log2(min_side));
 
-        // Iterate over patches
-        for (auto i = 0; i < 4; i++) {
-            auto& patch = *patches[i];
+        // // Iterate over patches
+        // for (auto i = 0; i < 4; i++) {
+        //     auto& patch = *patches[i];
 
-            // Iterate over number of tagged
-            for (auto n = 0; n < tags[i]; n++) {
-                auto& grid = patch.grid();
-                int ngrid = grid.nx();
-                int coarsen_factor = tags[i] - (tags[i] - n);
-                int nfine = ngrid / pow(2, coarsen_factor);
-                int ncoarse = nfine / 2;
+        //     // Iterate over number of tagged
+        //     for (auto n = 0; n < tags[i]; n++) {
+        //         auto& grid = patch.grid();
+        //         int ngrid = grid.nx();
+        //         int coarsen_factor = tags[i] - (tags[i] - n);
+        //         int nfine = ngrid / pow(2, coarsen_factor);
[…]
