### 2024-02-14 7b46b52
WIP: Debugging adaptive rebuild.
