### 2024-02-27 aa8737c
Adding matplotlibcpp capability back in for debugging.
