+++ src/PlotUtils.hpp
@@ -3,9 +3,11 @@
 
 #if USE_MATPLOTLIBCPP
 
+#include <utility>
 #include <matplotlibcpp.h>
 #include "Vector.hpp"
 #include "Matrix.hpp"
+#include "PatchGrid.hpp"
 
 namespace matplotlibcpp {
 
@@ -35,6 +37,34 @@ void matshow(EllipticForest::Matrix<NumericalType>& A, double tolerance = 1e-12)
         }
     }
     scatter(row, col, 1, {{"marker", "s"}, {"color", "k"}});
+
+}
+
+template<typename NumericalType>
+std::pair<EllipticForest::Vector<NumericalType>, EllipticForest::Vector<NumericalType>> meshgrid(EllipticForest::Vector<NumericalType>& x1, EllipticForest::Vector<NumericalType>& x2) {
+
+    auto M = x1.size();
+    auto N = x2.size();
+    EllipticForest::Vector<NumericalType> x1_mesh(M*N);
+    EllipticForest::Vector<NumericalType> x2_mesh(M*N);
+    for (auto i = 0; i < M; i++) {
+        for (auto j = 0; j < N; j++) {
+            auto ii = j + i*N;
+            x1_mesh[ii] = x1[i];
+            x2_mesh[ii] = x2[j];
+        }
+    }
+    return {x1_mesh, x2_mesh};
+
+}
+
+template<typename NumericalType>
+bool scatter3(EllipticForest::PatchGridBase<NumericalType>& grid, EllipticForest::Vector<NumericalType>& values) {
+
+    auto x1_points = grid.xPoints();
+    auto x2_points = grid.yPoints();
+    auto [x1_mesh, x2_mesh] = meshgrid(x1_points, x2_points);
+    scatter(x1_mesh.data(), x2_mesh.data(), values.data());
 
 }
 
