+++ config.sh
@@ -11,7 +11,13 @@ MPI_PATH=/opt/homebrew
 P4EST_PATH=${HOME}/packages/p4est/p4est/build/local
 
 # PETSC_PATH : Path to PETSc install (i.e., ${PETSC_PATH}/include, ${PETSC_PATH}/lib, ...) 
-PETSC_PATH=${HOME}/packages/petsc/petsc-build-mumps
+PETSC_PATH=/opt/homebrew
+
+# PYTHON_ENV_PATH : Path to Python conda environment that has numpy and matplotlib installed
+PYTHON_ENV_PATH=${HOME}/miniforge3/envs/EllipticForest
+
+# PYTHON_VERSION : Version of python in conda environment path
+PYTHON_VERSION=python3.11
 
 # BUILD_DIR : Build directory
 BUILD_DIR=build-$(git branch --show-current)
@@ -24,4 +30,6 @@ cmake -S . -B ${BUILD_DIR} \
     -DCMAKE_C_COMPILER=mpicc \
     -DMPI_PATH=${MPI_PATH} \
     -DP4EST_PATH=${P4EST_PATH} \
-    -DPETSC_PATH=${PETSC_PATH}
+    -DPETSC_PATH=${PETSC_PATH} \
+    -DPYTHON_ENV_PATH=${PYTHON_ENV_PATH} \
+    -DPYTHON_VERSION=${PYTHON_VERSION}
