### 2024-02-27 83abafe
More work on debugging adaptive rebuild.
