+++ test/test_interpolation.cpp
@@ -4,8 +4,14 @@
 
 using namespace EllipticForest;
 
+double f1d(double x) {
+    return sin(x) + 4.0;
+}
+
 double f2d(double x, double y) {
-    return x*x + y*y + 4.0;
+    return 30000.0*(sin(x) + pow(sin(y), 3));
+    // return x*x + y*y + 4.0;
+    // return x + y;
 }
 
 TEST(Interpolation, linear) {
@@ -38,6 +44,38 @@ TEST(Interpolation, linear) {
 
 }
 
+TEST(Interpolation, polynomial) {
+    
+    int n = 16;
+    Vector<double> x = linspace<double>(-2, 2, n);
+    Vector<double> y(n);
+    for (int i = 0; i < y.size(); i++) {
+        y[i] = f1d(x[i]);
+    }
+
+    PolynomialInterpolant f(x, y, 2);
+
+    double x_test = 0.5;
+    double y_test = f(x_test);
+    double y_true = f1d(x_test);
+    double error = fabs(y_test - y_true);
+
+    EXPECT_NEAR(y_test, y_true, 1e-2);
+
+    Vector<double> xv = linspace<double>(-2, 0, n);
+    Vector<double> yv_test = f(xv);
+    Vector<double> yv_true(n);
+    for (int i = 0; i < n; i++) {
+        yv_true[i] = f1d(xv[i]);
+        // printf("x = %f, y_test = %f, y_true = %f, error = %e\n", xv[i], yv_test[i], yv_true[i], fabs(yv_test[i] - yv_true[i]));
+    }
+
+    for (int i = 0; i < yv_true.size(); i++) {
+        EXPECT_NEAR(yv_test[i], yv_true[i], 1e-2);
+    }
+
+}
+
 TEST(Interpolation, bilinear) {
 
     double a = -1;
@@ -59,19 +97,102 @@ TEST(Interpolation, bilinear) {
 
     BilinearInterpolant I(x1_coarse, x2_coarse, y_coarse);
 
+    b = b / 2.;
+    d = d / 2.;
     Vector<double> x1_fine = linspace(a + dx/4, b - dx/4, m*2);
     Vector<double> x2_fine = linspace(c + dy/4, d - dy/4, n*2);
-    Vector<double> y_fine = I(x1_fine, x2_fine);
+    Vector<double> y_test = I(x1_fine, x2_fine);
 
-    Vector<double> y_fine_exp(4*m*n);
+    Vector<double> y_true(4*m*n);
+    Vector<double> rel_err(4*m*n);
     for (int i = 0; i < 2*m; i++) {
         for (int j = 0; j < 2*n; j++) {
-            y_fine_exp[j + 2*i*n] = f2d(x1_fine[i], x2_fine[j]);
+            int ii = j + 2*i*n;
+            y_true[ii] = f2d(x1_fine[i], x2_fine[j]);
+            rel_err[ii] = fabs(y_test[ii] - y_true[ii]) / y_true[ii];
+            // printf("x1 = %f, x2 = %f, y_test = %f, y_true = %f, error = %e\n", x1_fine[i], x2_fine[j], y_test[ii], y_true[ii], rel_err[ii]);
         }
     }
 
-    Vector<double> diff = y_fine_exp - y_fine;
-    double max_diff = vectorInfNorm(y_fine_exp, y_fine);
-    EXPECT_LT(max_diff, 1e-2);
+    double max_error = *std::max(rel_err.data().begin(), rel_err.data().end());
+    EXPECT_LT(max_error, 1e-2);
 
 }
+
+// TEST(Interpolation, polynomial_2d) {
+
+//     double a = 0;
+//     double b = 1;
+//     double c = 0;
[…]
