### 2024-02-29 8a79581
Added some quality of life features: more timers and options for elliptic-single.
