### 2024-03-09 cc71e1e
Adding strong scaling directory. Changed five-point stencil solver to sparse direct solver.
