+++ src/Patches/FiniteVolume/FiniteVolumeSolver.cpp
@@ -40,12 +40,10 @@ Vector<double> FiniteVolumeSolver::solve(FiniteVolumeGrid& grid, Vector<double>&
         double dy = grid.dy();
 
         // Get Petsc data
-        // ParallelMatrix<double> A(MPI_COMM_SELF, N, N, N, N, MATDENSE);
         Mat A;
         MatCreate(MPI_COMM_SELF, &A);
         MatSetSizes(A, N, N, N, N);
-        // MatSetFromOptions(A);
-        MatSetType(A, MATDENSE);
+        MatSetType(A, MATAIJ);
         MatSetUp(A);
 
         Vec f;
@@ -193,18 +191,17 @@ Vector<double> FiniteVolumeSolver::solve(FiniteVolumeGrid& grid, Vector<double>&
         MatAssemblyEnd(A, MAT_FINAL_ASSEMBLY);
         VecAssemblyEnd(f);
 
-        // Setup factorization
-        IS row_perm, col_perm;
-        Vector<int> idx_row = vectorRange(0, N-1);
-        Vector<int> idx_col = vectorRange(0, N-1);
-        ISCreateGeneral(MPI_COMM_WORLD, N, idx_row.data().data(), PETSC_COPY_VALUES, &row_perm);
-        ISCreateGeneral(MPI_COMM_WORLD, N, idx_col.data().data(), PETSC_COPY_VALUES, &col_perm);
-        MatFactorInfo mat_factor_info;
-        MatFactorInfoInitialize(&mat_factor_info);
+        // Setup KSP
+        KSP ksp;
+        PC pc;
+        KSPCreate(MPI_COMM_SELF, &ksp);
+        KSPSetType(ksp, KSPPREONLY);
+        KSPSetOperators(ksp, A, A);
+        KSPGetPC(ksp, &pc);
+        PCSetType(pc, PCLU);
 
-        // Solve linear system
-        MatLUFactor(A, row_perm, col_perm, &mat_factor_info);
-        MatSolve(A, f, x);
+        // Solve system
+        KSPSolve(ksp, f, x);
 
         // Create EllipticForest vector
         double* x_data;
@@ -217,7 +214,7 @@ Vector<double> FiniteVolumeSolver::solve(FiniteVolumeGrid& grid, Vector<double>&
         MatDestroy(&A);
         VecDestroy(&x);
         VecDestroy(&f);
-        // KSPDestroy(&ksp);
+        KSPDestroy(&ksp);
 
         return x_vector;
     }
