+++ examples/strong-scaling/main.cpp
@@ -1,0 +1,298 @@
+/**
+ * @file main.cpp : elliptic-single
+ * @author Damyn Chipman (DamynChipman@u.boisestate.edu)
+ * @brief Sets up and solves an elliptic PDE using the Hierarchical Poincaré-Steklov (HPS) method.
+ * 
+ * Solves Poisson's equation:
+ * 
+ * laplacian( u ) = f
+ * 
+ * subject to Dirichlet boundary conditions provided by the exact solution.
+ * 
+ * By default, this is set to solve for the exact solution:
+ * 
+ * u(x,y) = sin(x) + sin(y)
+ * 
+ * thus,
+ * 
+ * f(x,y) = -sin(x) - sin(y) = -u(x,y).
+ * 
+ * EllipticForest solves this by creating a mesh and refining it according to the curvature of the
+ * solution (i.e., the right-hand side function `f`). The build, upwards, and solve stages are used
+ * to do the factorization and application of the solution operators. The solution is output to VTK
+ * files to be viewed with your favorite visualization tool (VisIt is mine!)
+ * 
+ */
+
+#include <cmath>
+#include <iostream>
+#include <string>
+
+#include <EllipticForest.hpp>
+#include <Patches/FiniteVolume/FiniteVolume.hpp>
+
+/**
+ * @brief User-defined solution function
+ * 
+ * When computing error, calls this function to compare with numerical solution.
+ * 
+ * Used by EllipticForest to create the boundary data.
+ * 
+ * @param x x-coordinate
+ * @param y y-coordinate
+ * @return double 
+ */
+double uFunction(double x, double y) {
+    return sin(x) + sin(y);
+}
+
+/**
+ * @brief User-defined load function
+ * 
+ * Right-hand side function of the elliptic PDE.
+ * 
+ * @param x x-coordinate
+ * @param y y-coordinate
+ * @return double 
+ */
+double fFunction(double x, double y) {
+    return -uFunction(x, y);
+}
+
+/**
+ * @brief User-defined alpha function
+ * 
+ * @param x x-coordinate
+ * @param y y-coordinate
+ * @return double 
+ */
+double alphaFunction(double x, double y) {
+    return 1.0;
+}
+
+/**
+ * @brief User-defined beta function
+ * 
+ * @param x x-coordinate
+ * @param y y-coordinate
+ * @return double 
+ */
+double betaFunction(double x, double y) {
+    return 1.0;
+}
+
+/**
+ * @brief User-defined lambda function
+ * 
+ * @param x x-coordinate
+ * @param y y-coordinate
+ * @return double 
+ */
+double lambdaFunction(double x, double y) {
+    return 0.0;
+}
+
+/**
+ * @brief Main driver for elliptic-single
+ * 
+ * Solves an elliptic PDE with user defined setup functions using EllipticForest
+ * 
+ * @param argc 
+ * @param argv 
+ * @return int 
+ */
+int main(int argc, char** argv) {
+
+    // ====================================================
+    // Initialize app and MPI
+    // ====================================================
+    EllipticForest::EllipticForestApp app(&argc, &argv);
+    EllipticForest::MPI::MPIObject mpi(MPI_COMM_WORLD);
+
+    // ====================================================
+    // Setup options
+    // ====================================================
+    bool cache_operators = false;
+    app.options.setOption("cache-operators", cache_operators);
+
+    bool homogeneous_rhs = false;
+    app.options.setOption("homogeneous-rhs", homogeneous_rhs);
+
+    bool vtk_flag = true;
+    app.options.setOption("vtk-flag", vtk_flag);
+    
+    double threshold = 1.2;
[…]
