+++ examples/CMakeLists.txt
@@ -3,5 +3,6 @@ add_subdirectory(hello)
 add_subdirectory(patch-solver)
 add_subdirectory(elliptic-single)
 add_subdirectory(elliptic-multiple)
+add_subdirectory(strong-scaling)
 add_subdirectory(thermal)
 # add_subdirectory(papers/p1-serial)
