### 2024-03-12 3a0a45e
Adding strong and weak scaling examples.
