+++ examples/weak-scaling/main.cpp
@@ -1,5 +1,5 @@
 /**
- * @file main.cpp : elliptic-single
+ * @file main.cpp : weak-scaling
  * @author Damyn Chipman (DamynChipman@u.boisestate.edu)
  * @brief Sets up and solves an elliptic PDE using the Hierarchical Poincaré-Steklov (HPS) method.
  * 
@@ -27,37 +27,138 @@
 #include <cmath>
 #include <iostream>
 #include <string>
+#include <algorithm>
 
 #include <EllipticForest.hpp>
 #include <Patches/FiniteVolume/FiniteVolume.hpp>
 
-/**
- * @brief User-defined solution function
- * 
- * When computing error, calls this function to compare with numerical solution.
- * 
- * Used by EllipticForest to create the boundary data.
- * 
- * @param x x-coordinate
- * @param y y-coordinate
- * @return double 
- */
-double uFunction(double x, double y) {
-    return sin(x) + sin(y);
-}
+class PolarStarPoissonProblem {
 
-/**
- * @brief User-defined load function
- * 
- * Right-hand side function of the elliptic PDE.
- * 
- * @param x x-coordinate
- * @param y y-coordinate
- * @return double 
- */
-double fFunction(double x, double y) {
-    return -uFunction(x, y);
-}
+public:
+
+    int nPolar;
+    std::vector<double> x0s;
+    std::vector<double> y0s;
+    std::vector<double> r0s;
+    std::vector<double> r1s;
+    std::vector<double> ns;
+    double eps_disk = 0.015625;
+
+    PolarStarPoissonProblem() {
+        nPolar = 1;
+        x0s = {0};
+        y0s = {0};
+        r0s = {0.4};
+        r1s = {0.45};
+        ns = {4};
+    }
+
+    PolarStarPoissonProblem(int nPolar, std::vector<double> x0s, std::vector<double> y0s, std::vector<double> r0s, std::vector<double> r1s, std::vector<double> ns, double epsilon) :
+        nPolar(nPolar),
+        x0s(x0s),
+        y0s(y0s),
+        r0s(r0s),
+        r1s(r1s),
+        ns(ns),
+        eps_disk(epsilon)
+            {}
+
+    std::string name() { return "polar-star-poisson"; }
+
+    double lambda() { return 0.0; }
+
+    double u(double x, double y) {
+        double res = 0;
+        for (auto i = 0; i < nPolar; i++) {
+            double x0 = x0s[i];
+            double y0 = y0s[i];
+            double r = sqrt(pow(x - x0, 2) + pow(y - y0, 2));
+            double theta = atan2(y - y0, x - x0);
+            res += 1.0 - hsmooth(i, r, theta);
+        }
+        return res;
+    }
+
+    double f(double x, double y) {
+        double res = 0;
+        for (auto i = 0; i < nPolar; i++) {
+            double x0 = x0s[i];
+            double y0 = y0s[i];
+            double r = sqrt(pow(x - x0, 2) + pow(y - y0, 2));
+            double theta = atan2(y - y0, x - x0);
+            res -= hsmooth_laplacian(i, r, theta);
+        }
+        return res;
+    }
+
+    double dudx(double x, double y) {
+        return 0;
+    }
+
+    double dudy(double x, double y) {
+        return 0;
+    }
+
+private:
+
+    double sech(double x) {
+        return 1.0 / cosh(x);
+    }
+
+    void polar_interface_complete(int ID, double theta, double& p, double& dpdtheta, double& d2pdtheta2) {
+        double r0 = r0s[ID];
+        double r1 = r1s[ID];
+        int n = ns[ID];
+
+        p = r0*(1.0 + r1*cos(n*theta));
+        dpdtheta = r0*(-n*r1*sin(n*theta));
[…]
