### 2024-04-03 5d2dce6
WIP: Tinkering with adaptive rebuild still. Sanity checks in toybox: compare refined mesh to refactored mesh.
