+++ src/Patches/FiniteVolume/FiniteVolumeNodeFactory.cpp
@@ -66,99 +66,129 @@ Node<FiniteVolumePatch>* FiniteVolumeNodeFactory::createChildNode(Node<FiniteVol
         tagged_for_refinement = true;
     }
 
-    if (parent_patch.vectorG().size() != 0) {
-        // app.log("[createChildNode] creating child g");
-        auto parent_x_W = linspace(parent_grid(1, 0), parent_grid(1, parent_grid.ny()-1), parent_grid.ny());
-        auto parent_x_E = linspace(parent_grid(1, 0), parent_grid(1, parent_grid.ny()-1), parent_grid.ny());
-        auto parent_x_S = linspace(parent_grid(0, 0), parent_grid(0, parent_grid.nx()-1), parent_grid.nx());
-        auto parent_x_N = linspace(parent_grid(0, 0), parent_grid(0, parent_grid.nx()-1), parent_grid.nx());
+//     if (parent_patch.vectorG().size() != 0) {
+//         // app.log("[createChildNode] creating child g");
+        
+//         int n_parent = parent_grid.nx();
+//         double h_child = child_grid.dx();
+//         int n_child = child_grid.nx();
+//         double h_parent = parent_grid.dx();
 
-        auto parent_I_W = vectorRange(0, parent_grid.ny() - 1);
-        auto parent_I_E = vectorRange(parent_grid.ny(), 2*parent_grid.ny() - 1);
-        auto parent_I_S = vectorRange(2*parent_grid.ny(), (2*parent_grid.ny() + parent_grid.nx()) - 1);
-        auto parent_I_N = vectorRange(2*parent_grid.ny() + parent_grid.nx(), (2*parent_grid.ny() + 2*parent_grid.nx()) - 1);
+//         auto parent_y = linspace(parent_grid.yLower() + h_parent/2, parent_grid.yUpper() - h_parent/2, n_parent);
+//         auto parent_x = linspace(parent_grid.xLower() + h_parent/2, parent_grid.xUpper() - h_parent/2, n_parent);
 
-        auto parent_g_W = parent_patch.vectorG()(parent_I_W);
-        auto parent_g_E = parent_patch.vectorG()(parent_I_E);
-        auto parent_g_S = parent_patch.vectorG()(parent_I_S);
-        auto parent_g_N = parent_patch.vectorG()(parent_I_N);
+//         // auto parent_x_W = linspace(parent_grid(1, 0), parent_grid(1, parent_grid.ny()-1), parent_grid.ny());
+//         // auto parent_x_E = linspace(parent_grid(1, 0), parent_grid(1, parent_grid.ny()-1), parent_grid.ny());
+//         // auto parent_x_S = linspace(parent_grid(0, 0), parent_grid(0, parent_grid.nx()-1), parent_grid.nx());
+//         // auto parent_x_N = linspace(parent_grid(0, 0), parent_grid(0, parent_grid.nx()-1), parent_grid.nx());
 
-        LinearInterpolant interpolant_W(parent_x_W, parent_g_W);
-        LinearInterpolant interpolant_E(parent_x_E, parent_g_E);
-        LinearInterpolant interpolant_S(parent_x_S, parent_g_S);
-        LinearInterpolant interpolant_N(parent_x_N, parent_g_N);
+//         // auto parent_I_W = vectorRange(0, parent_grid.ny() - 1);
+//         // auto parent_I_E = vectorRange(parent_grid.ny(), 2*parent_grid.ny() - 1);
+//         // auto parent_I_S = vectorRange(2*parent_grid.ny(), (2*parent_grid.ny() + parent_grid.nx()) - 1);
+//         // auto parent_I_N = vectorRange(2*parent_grid.ny() + parent_grid.nx(), (2*parent_grid.ny() + 2*parent_grid.nx()) - 1);
 
[…]
