+++ src/HPSAlgorithm.hpp
@@ -14,6 +14,7 @@
 #include "Matrix.hpp"
 #include "SpecialMatrices.hpp"
 #include "Mesh.hpp"
+#include "Interpolation.hpp"
 
 namespace EllipticForest {
 
@@ -150,7 +151,7 @@ public:
                 PatchType& beta = child_nodes[1]->data;
                 PatchType& gamma = child_nodes[2]->data;
                 PatchType& omega = child_nodes[3]->data;
-                merge4to1(tau, alpha, beta, gamma, omega);
+                merge4to1(tau, alpha, beta, gamma, omega, patch_solver);
                 return 1;
             }
         );
@@ -278,6 +279,7 @@ public:
                 PatchType& gamma = child_nodes[2]->data;
                 PatchType& omega = child_nodes[3]->data;
                 upwards4to1(tau, alpha, beta, gamma, omega);
+                // app.log("w: [%6i]", parent_node->data.vectorW().size());
                 return 1;
             }
         );
@@ -452,6 +454,7 @@ public:
             },
             [&](Node<PatchType>* parent_node, std::vector<Node<PatchType>*> child_nodes){
                 // app.log("Family callback: parent path = " + parent_node->path);
+                // app.log("w: [%6i]", parent_node->data.vectorW().size());
                 // Family callback
                 PatchType& tau = parent_node->data;
                 PatchType& alpha = child_nodes[0]->data;
@@ -477,7 +480,7 @@ public:
      * @param gamma Upper-left patch
      * @param omega Upper-right patch
      */
-    static void merge4to1(PatchType& tau, PatchType& alpha, PatchType& beta, PatchType& gamma, PatchType& omega) {
+    static void merge4to1(PatchType& tau, PatchType& alpha, PatchType& beta, PatchType& gamma, PatchType& omega, PatchSolverType& patch_solver) {
         
         EllipticForestApp& app = EllipticForestApp::getInstance();
         // app.logHead("Merging:");
@@ -536,14 +539,14 @@ public:
         //         auto& grid = patch.grid();
         //         int ngrid = grid.nx();
         //         int coarsen_factor = tags[i] - (tags[i] - n);
-        //         int nfine = ngrid / pow(2, coarsen_factor);
-        //         int ncoarse = nfine / 2;
+        //         int n_fine = ngrid / pow(2, coarsen_factor);
+        //         int ncoarse = n_fine / 2;
 
         //         InterpolationMatrixFine2Coarse<NumericalType> L21_side(ncoarse);
         //         std::vector<Matrix<NumericalType>> L21_diagonals = {L21_side, L21_side, L21_side, L21_side};
         //         Matrix<NumericalType> L21_patch = blockDiagonalMatrix(L21_diagonals);
 
-        //         InterpolationMatrixCoarse2Fine<NumericalType> L12_side(nfine);
+        //         InterpolationMatrixCoarse2Fine<NumericalType> L12_side(n_fine);
         //         std::vector<Matrix<NumericalType>> L12_diagonals = {L12_side, L12_side, L12_side, L12_side};
         //         Matrix<NumericalType> L12_patch = blockDiagonalMatrix(L12_diagonals);
 
@@ -556,6 +559,7 @@ public:
 
         // Coarsen children patches to match siblings
         std::vector<PatchType*> patches = {&alpha, &beta, &gamma, &omega};
+
         std::vector<std::size_t> sizes = {
             alpha.grid().nx(),
             beta.grid().nx(),
@@ -566,31 +570,46 @@ public:
         PatchGridType merged_grid(MPI_COMM_SELF, 2*min_size, alpha.grid().xLower(), beta.grid().xUpper(), 2*min_size, alpha.grid().yLower(), gamma.grid().yUpper());
         tau.grid() = merged_grid;
 
[…]
