### 2024-04-05 d597b29
Adding Helmholtz example.
