+++ examples/helmholtz/main.cpp
@@ -1,0 +1,260 @@
+/**
+ * @file main.cpp : elliptic-single
+ * @author Damyn Chipman (DamynChipman@u.boisestate.edu)
+ * @brief Sets up and solves an elliptic PDE using the Hierarchical Poincaré-Steklov (HPS) method.
+ * 
+ * Solves Poisson's equation:
+ * 
+ * laplacian( u ) = f
+ * 
+ * subject to Dirichlet boundary conditions provided by the exact solution.
+ * 
+ * By default, this is set to solve for the exact solution:
+ * 
+ * u(x,y) = sin(x) + sin(y)
+ * 
+ * thus,
+ * 
+ * f(x,y) = -sin(x) - sin(y) = -u(x,y).
+ * 
+ * EllipticForest solves this by creating a mesh and refining it according to the curvature of the
+ * solution (i.e., the right-hand side function `f`). The build, upwards, and solve stages are used
+ * to do the factorization and application of the solution operators. The solution is output to VTK
+ * files to be viewed with your favorite visualization tool (VisIt is mine!)
+ * 
+ */
+
+#include <cmath>
+#include <iostream>
+#include <string>
+
+#include <EllipticForest.hpp>
+#include <Patches/FiniteVolume/FiniteVolume.hpp>
+
+/**
+ * @brief Wrapper of cmath Bessel Y function
+ * 
+ * @param n Order of Bessel function
+ * @param x Value of x
+ * @return double 
+ */
+double besselY(int n, double x) {
+    return yn(n, x);
+}
+
+/**
+ * @brief Main driver for elliptic-single
+ * 
+ * Solves an elliptic PDE with user defined setup functions using EllipticForest
+ * 
+ * @param argc 
+ * @param argv 
+ * @return int 
+ */
+int main(int argc, char** argv) {
+
+    // ====================================================
+    // Initialize app and MPI
+    // ====================================================
+    EllipticForest::EllipticForestApp app(&argc, &argv);
+    EllipticForest::MPI::MPIObject mpi(MPI_COMM_WORLD);
+
+    // ====================================================
+    // Setup options
+    // ====================================================
+    bool cache_operators = false;
+    app.options.setOption("cache-operators", cache_operators);
+
+    bool homogeneous_rhs = false;
+    app.options.setOption("homogeneous-rhs", homogeneous_rhs);
+
+    bool vtk_flag = true;
+    app.options.setOption("vtk-flag", vtk_flag);
+    
+    double threshold = 1.2;
+    app.options.setOption("refinement-threshold", threshold);
+    
+    int n_solves = 1;
+    app.options.setOption("n-solves", n_solves);
+    
+    double x_lower = -10.0;
+    app.options.setOption("x-lower", x_lower);
+
+    double x_upper = 10.0;
+    app.options.setOption("x-upper", x_upper);
+
+    double y_lower = -10.0;
+    app.options.setOption("y-lower", y_lower);
+
+    double y_upper = 10.0;
+    app.options.setOption("y-upper", y_upper);
+    
+    int min_level = 0;
+    int max_level = 5;
+    int nx = 16;
+    int ny = 16;
+    if (argc > 1) {
+        min_level = atoi(argv[1]);
+        max_level = atoi(argv[2]);
+        nx = atoi(argv[3]);
+        ny = atoi(argv[4]);
+    }
+    app.options.setOption("max-level", max_level);
+    app.options.setOption("min-level", min_level);
+    app.options.setOption("nx", nx);
+    app.options.setOption("ny", ny);
+
+    double lambda = 100.;
+    double kappa = sqrt(lambda);
+    double x0 = -20.;
+    double y0 = 0.;
+    double amplitude = 1.;
+    double sigma_x = 0.1;
+    double sigma_y = 0.4;
+
[…]
