+++ examples/CMakeLists.txt
@@ -6,4 +6,5 @@ add_subdirectory(elliptic-multiple)
 add_subdirectory(strong-scaling)
 add_subdirectory(weak-scaling)
 add_subdirectory(thermal)
+add_subdirectory(helmholtz)
 # add_subdirectory(papers/p1-serial)
