### 2024-04-09 26e476b
Updating Helmholtz with proper documentation and output.
