### 2024-04-16 DamynChipman/EllipticForest#8 @lukeolson
joss paper suggestions
https://github.com/openjournals/joss-reviews/issues/6339

Here are a few suggestions for the paper:
- line 14: fix "found the FISHPACK"
- line 29: makes -> make
- line 30: "Further complicating implementations of effective solvers" -> "In addition" (the readability of this sentence could be improved)
- line 53: does "tossed" mean the factorization itself?  You may get by with just using "recomputed" here.
- In the reference, please consider adding DOI information for the journal articles
- Lines 158/160: are these "Forthcoming" articles available as preprints?  If not then I suggest removing them (or ask the editor what the standard is for this).
- Line 161: Poincaré and Steklov should be capitalized
- Line 166: This is a SIAM SISC article -- the bibtex is
```bibtex
@article{doi:10.1137/130918988,
    author = {Gillman, A. and Martinsson, P. G.},
    title = {A Direct Solver with \$O(N)\$ Complexity for Variable Coefficient Elliptic PDEs Discretized via a High-Order Composite Spectral Collocation Method},
    journal = {SIAM Journal on Scientific Computing},
    volume = {36},
    number = {4},
    pages = {A2023-A2046},
    year = {2014},
    doi = {10.1137/130918988},
    url = {https://doi.org/10.1137/130918988},
    eprint = {https://doi.org/10.1137/130918988},
}

```
- Line 169: Poincaré and Steklov should be capitalized
- Line 177: there may be a more standard reference for MPI
