### 2024-04-17 DamynChipman/EllipticForest#9 @lukeolson
joss paper example
I see that the JOSS paper references an elliptic polar star example: `examples/elliptic-single/output_polar_star.png`.  Is there a way to reproduce this result in `examples`?  `v1-joss-paper` does not seem to have this in `examples/elliptic-single` but I may be missing something...
