### 2024-04-17 #6339 @lukeolson
@vissarion I've finished a pass at everything and left two comments in the form of issues for @DamynChipman .  The first is a list of suggestions for revising the write up.  The second is a reproducibility question -- I would like to see if it's possible to reproduce the figure in the paper.  Other than that, everything check on my end.
