### 2024-04-17 DamynChipman/EllipticForest#8 @DamynChipman
Awesome, thanks for the review. I will look at each of these over the next couple of days.
