### 2024-04-17 DamynChipman/EllipticForest#9 @DamynChipman
Ah, you're right. The polar-star example can be found in the main branch, at `examples/papers/p1-serial/poisson-2.cpp` (https://github.com/DamynChipman/EllipticForest/blob/main/examples/papers/p1-serial/poisson-2.cpp). It's an example of a Poisson equation with a highly local RHS function. The `v1-joss-paper` branch `elliptic-single` example is set up to solve the Poisson equation for $u(x,y) = \sin(x) + \sin(y)$. I put the polar star example in the paper as it highlights the adaptive mesh better than the sine problem.

You can run the examples in `examples/papers/p1-serial` to reproduce those results. They correspond to the examples ran for a different paper, but they should work in parallel regardless.

Alternatively, I could either change the picture in the paper to match the example in `elliptic-single` or change the problem solved in `elliptic-single` to be the polar star problem. Any suggestions?
