+++ paper.bib
@@ -1,53 +1,61 @@
-@article{chipman2024serial,
-  author          = {Chipman, D and Calhoun, D and Burstedde, C},
-  journal         = {[Forthcoming]},
-  title           = {A Fast Direct Solver for Elliptic PDEs on a Hierarchy of Adaptively Refined Quadtrees},
-  year            = {2024}
+@article{chipman2024fast,
+	title={A Fast Direct Solver for Elliptic PDEs on a Hierarchy of Adaptively Refined Quadtrees},
+	author={Chipman, Damyn and Calhoun, Donna and Burstedde, Carsten},
+	journal={arXiv preprint arXiv:2402.14936},
+	year={2024},
+	doi={https://doi.org/10.48550/arXiv.2402.14936}
 }
 
 @article{chipman2024parallel,
-  author          = {Chipman, D and Calhoun, D and Burstedde, C},
-  journal         = {[Forthcoming]},
-  title           = {Parallel and Adaptive Optimizations to the Quadtree-Adaptive Hierarchical Poincaré-Steklov Method},
-  year            = {2024}
+	author = {Chipman, D and Calhoun, D and Burstedde, C},
+	journal = {[Forthcoming]},
+	title = {{MPI Implementation of the Quadtree-Adaptive Hierarchical Poincar{\'e}-Steklov Method}},
+	year = {2024}
 }
 
 @article{gillman2014direct,
-	author = {Gillman, A and Martinsson, PG},
-	journal = siam_sc,
-	number = {4},
-	pages = {A2023--A2046},
-	title = {{A direct solver with O(N) complexity for variable coefficient elliptic PDEs discretized via a high-order composite spectral collocation method}},
-	volume = {36},
-	year = {2014},
-	doi = {10.1137/130918988}}
+    author = {Gillman, A. and Martinsson, P. G.},
+    title = {A Direct Solver with \$O(N)\$ Complexity for Variable Coefficient Elliptic PDEs Discretized via a High-Order Composite Spectral Collocation Method},
+    journal = {SIAM Journal on Scientific Computing},
+    volume = {36},
+    number = {4},
+    pages = {A2023-A2046},
+    year = {2014},
+    doi = {10.1137/130918988},
+    url = {https://doi.org/10.1137/130918988},
+    eprint = {https://doi.org/10.1137/130918988},
+}
 
 @book{martinsson2019fast,
 	author = {Martinsson, PG},
 	publisher = {SIAM},
 	title = {{Fast direct solvers for elliptic PDEs}},
-	year = {2019}}
+	year = {2019},
+	doi={https://doi.org/10.1137/1.9781611976045}
+}
 
 @article{swarztrauber1999fishpack,
 	author = {Swarztrauber, P and Sweet, R and Adams, J},
 	journal = {UCAR Publication, July},
 	title = {FISHPACK: Efficient FORTRAN subprograms for the solution of elliptic partial differential equations},
-	year = {1999}}
+	year = {1999}
+}
 
 @article{adams2016fishpack90,
 	author = {Adams, J C and Swarztrauber, P and Sweet, R},
 	journal = {Astrophysics Source Code Library},
 	pages = {ascl--1609},
 	title = {FISHPACK90: Efficient fortran subprograms for the solution of separable elliptic partial differential equations},
-	year = {2016}}
+	year = {2016}
+}
 
-@article{walker1996mpi,
-	author = {Walker, D and Dongarra, J},
-	journal = {Supercomputer},
-	pages = {56--68},
-	title = {MPI: a standard message passing interface},
[…]
