+++ paper.bib
@@ -1,9 +1,9 @@
 @article{chipman2024fast,
 	title={A Fast Direct Solver for Elliptic PDEs on a Hierarchy of Adaptively Refined Quadtrees},
-	author={Chipman, Damyn and Calhoun, Donna and Burstedde, Carsten},
+	author={Chipman, D and Calhoun, D and Burstedde, C},
 	journal={arXiv preprint arXiv:2402.14936},
 	year={2024},
-	doi={https://doi.org/10.48550/arXiv.2402.14936}
+	doi={10.48550/arXiv.2402.14936}
 }
 
 @article{chipman2024parallel,
@@ -31,7 +31,7 @@
 	publisher = {SIAM},
 	title = {{Fast direct solvers for elliptic PDEs}},
 	year = {2019},
-	doi={https://doi.org/10.1137/1.9781611976045}
+	doi={10.1137/1.9781611976045}
 }
 
 @article{swarztrauber1999fishpack,
@@ -65,7 +65,7 @@
 	title = {{\texttt{p4est}}: Scalable Algorithms for Parallel Adaptive Mesh Refinement on Forests of Octrees},
 	volume = {33},
 	year = {2011},
-	doi={https://doi.org/10.1137/100791634}
+	doi={10.1137/100791634}
 }
 
 @incollection{paraview,
@@ -90,7 +90,7 @@
 	journal = {arXiv preprint arXiv:1506.01308},
 	title = {{The hierarchical Poincar{\'e}-Steklov (HPS) solver for elliptic PDEs: A tutorial}},
 	year = {2015},
-	doi={https://doi.org/10.48550/arXiv.1506.01308}
+	doi={10.48550/arXiv.1506.01308}
 }
 
 @incollection{quarteroni1991theory,
