### 2024-04-20 80b8886
Fixing LaTeX formatting.
