+++ paper.bib
@@ -38,7 +38,8 @@
 	author = {Swarztrauber, P and Sweet, R and Adams, J},
 	journal = {UCAR Publication, July},
 	title = {FISHPACK: Efficient FORTRAN subprograms for the solution of elliptic partial differential equations},
-	year = {1999}
+	year = {1999},
+	doi={10.1145/800207.806417}
 }
 
 @article{adams2016fishpack90,
@@ -46,7 +47,8 @@
 	journal = {Astrophysics Source Code Library},
 	pages = {ascl--1609},
 	title = {FISHPACK90: Efficient fortran subprograms for the solution of separable elliptic partial differential equations},
-	year = {2016}
+	year = {2016},
+	url={https://ui.adsabs.harvard.edu/abs/2016ascl.soft09005A/abstract}
 }
 
 @manual{mpi41,
@@ -77,12 +79,20 @@
 	doi = {10.1016/b978-012387582-2/50038-1}
 }
 
-@techreport{anl2023petsc,
-	author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed Brown and Peter Brune and Kris Buschelman and Emil Constantinescu and Lisandro Dalcin and Alp Dener and Victor Eijkhout and Jacob Faibussowitsch and William~D. Gropp and V\'{a}clav Hapla and Tobin Isaac and Pierre Jolivet and Dmitry Karpeev and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott Kruger and Dave~A. May and Lois Curfman McInnes and Richard Tran Mills and Lawrence Mitchell and Todd Munson and Jose~E. Roman and Karl Rupp and Patrick Sanan and Jason Sarich and Barry~F. Smith and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao Zhang},
-	institution = {Argonne National Laboratory},
-	number = {ANL-21/39 - Revision 3.20},
-	title = {{PETSc/TAO} Users Manual},
-	year = {2023}
+@Misc{            anl2023petsc,
+  author        = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed
+                  Brown and Peter Brune and Kris Buschelman and Emil~M. Constantinescu and Lisandro
+                  Dalcin and Alp Dener and Victor Eijkhout and Jacob Faibussowitsch and William~D.
+                  Gropp and V\'{a}clav Hapla and Tobin Isaac and Pierre Jolivet and Dmitry Karpeev
+                  and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott Kruger and
+                  Dave~A. May and Lois Curfman McInnes and Richard Tran Mills and Lawrence Mitchell
+                  and Todd Munson and Jose~E. Roman and Karl Rupp and Patrick Sanan and Jason Sarich
+                  and Barry~F. Smith and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao
+                  Zhang},
+  title         = {{PETS}c {W}eb page},
+  url           = {https://petsc.org/},
+  howpublished  = {\url{https://petsc.org/}},
+  year          = {2024}
 }
 
 @article{martinsson2015hierarchical,
@@ -109,7 +119,8 @@
 	month = {October},
 	pages = {357-372},
 	title = {VisIt: An End-User Tool For Visualizing and Analyzing Very Large Data},
-	year = {2012}
+	year = {2012},
+	doi={10.1201/b12985-29}
 }
 
 @misc{ultraSEM,
