### 2024-04-20 d796f29
Removing parallel citation to HPS method.
