### 2024-04-20 DamynChipman/EllipticForest#8 @DamynChipman
I incorporated your feedback into the paper, including fixes to the references. I'll update the draft on the JOSS review thread.

https://github.com/DamynChipman/EllipticForest/commit/d796f29d6f95873bcb615194ef33fe413030b4ee
