### 2024-04-20 #6339 @DamynChipman
The latest commit https://github.com/DamynChipman/EllipticForest/commit/d796f29d6f95873bcb615194ef33fe413030b4ee of the `v1-joss-paper` branch has the paper feedback from @lukeolson.
