### 2024-04-20 629f80d
Switching example picture to sine/cosine problem.
