### 2024-04-20 DamynChipman/EllipticForest#9 @DamynChipman
I went ahead and put the sine/cosine problem in the paper as it better corresponds to the current example in `examples/elliptic-single`. You can reproduce that example by following the instructions in the README: https://github.com/DamynChipman/EllipticForest/tree/v1-joss-paper/examples/elliptic-single. Each of the examples have a corresponding README with details about the example and how to run it.

The polar-star problem is also still available if you want to reproduce that in the directory I mentioned above.
