### 2024-04-20 #6339 @DamynChipman
@lukeolson, I addressed both of the issues you raised in the paper repo:

https://github.com/DamynChipman/EllipticForest/issues/8
https://github.com/DamynChipman/EllipticForest/issues/9
