+++ paper.bib
@@ -91,7 +91,6 @@
                   Zhang},
   title         = {{PETS}c {W}eb page},
   url           = {https://petsc.org/},
-  howpublished  = {\url{https://petsc.org/}},
   year          = {2024}
 }
 
