+++ paper.bib
@@ -15,7 +15,7 @@
 
 @article{gillman2014direct,
     author = {Gillman, A. and Martinsson, P. G.},
-    title = {{A Direct Solver with \$O(N)\$ Complexity for Variable Coefficient Elliptic PDEs Discretized via a High-Order Composite Spectral Collocation Method}},
+    title = {{A Direct Solver with $O(N)$ Complexity for Variable Coefficient Elliptic PDEs Discretized via a High-Order Composite Spectral Collocation Method}},
     journal = {SIAM Journal on Scientific Computing},
     volume = {36},
     number = {4},
