### 2024-04-21 #6339 @lukeolson
@DamynChipman the paper updates look great!  thank you for the thorough updates to the `.bib`!

The updates to the examples look good too.   I would encourage you to weave that polar star example into a separate example -- but for the paper everything looks good!
