### 2024-04-21 DamynChipman/EllipticForest#8 @lukeolson
looks good @DamynChipman
