### 2024-04-21 DamynChipman/EllipticForest#9 @lukeolson
@DamynChipman looks good!
