### 2024-04-22 #6339 @sandeshkatakam
> > > @DamynChipman thanks for the follow up.
> > > @sandeshkatakam, @lukeolson any news on this review?
> >
> > Hey I was travelling outside the country last week and couldn't respond
> >
> > I'll finish the review by the end of this week
> >
> > I'll let you know if i need something Thank you
>
> Hi, @sandeshkatakam it seems that your checklist is complete. Are you waiting for some feedback from the author, or you are done?

Yes I am done with my review. Good to go
