### 2024-04-22 #6339 @DamynChipman
@vissarion, done!

v1.0.2: https://github.com/DamynChipman/EllipticForest/releases/tag/v1.0.2
DOI: [10.5281/zenodo.11035127](https://doi.org/10.5281/zenodo.11035127)
