### 2024-04-25 #6339 @danielskatz
@DamynChipman - I've suggested some small changes in https://github.com/DamynChipman/EllipticForest/pull/11 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication
