### 2024-04-26 #6339 @danielskatz
Congratulations to @DamynChipman (Damyn Chipman) on your publication!!

JOSS depends on volunteers and couldn't succeed without you
